Re-export parent properties if called with a syspath
If we call this tool with a syspath argument, run up the udev
device parents to search for the various properties we set. If we find
any, re-export them.

In the udev rule we can then call this tool for each matching syspath
and will just copy the udev property up instead of re-running the usb
command.

Finally, in the udev rule we can copy the value over to UNIQ and
HID_UNIQ depending on the device we have.

No, for some reason IMPORT{parent} did not work for the HUION properties
we're adding...