udev: copy the HUION_FIRMWARE_ID into (HID_)UNIQ
libwacom relies on UNIQ to be parse the tools firmware id, so let's make
sure we copy that over. This now roughly matches the behaviour of the
DIGIMEND driver.

Since huion-switcher now copies the properties we can run it safely
enough on all actions - we need to run it on add now too to make sure
our properties get set correctly.

This was mentioned in 8b75cacbcb8b but somehow got lost in a rebase.

Fixes: 8b75cacbcb8b ("Re-export parent properties if called with a syspath")