fix re-exporting udev properties
According to the doc:
```
any() is short-circuiting; in other words, it will stop processing as
soon as it finds a true, given that no matter what else happens, the
result will also be true.
```

Which means that re-issuing the command (on udev `bind`) will stop after
printing the first matching property, effectively removing all of the
others.