Read the string descriptor for the HUION_MAGIC_BYTES manually
rusb's read_string_descriptor() enforces UTF16 with a peculiar check:

https://docs.rs/rusb/latest/src/rusb/device_handle.rs.html#757
    if len < 2 || buf[0] != len as u8 || len & 0x01 != 0 {
                    return Err(Error::BadDescriptor);
    }

This makes sense since it expects UTF16 but the Kamvas 12 returns a
buffer length is 19 - one byte for the length, one byte by the device
type (0x3) followed by 17 data bytes.  But we never get past this check
because rusb expects an even buffer length.
On the Huion Inspiroy 2S the total buffer is 20, so it works there.

So split this up: for the two queries that result in the magic bytes
being printed (index 200 and 100) use a custom message that returns the
full buffer including the two header bytes - this matches the
hid-uclogic driver which does all the checking on bytes 2+.

For the firmware and the pad mode use a normal string descriptor read,
just like hid-uclogic does.